refactor(gdocs): move notifications to own file
